requirements.txt
# Pharma Agentic AI - Dependencies
# CrewAI Framework
crewai>=0.36.0
crewai-tools>=0.4.0

# LLM Provider
groq>=0.4.0
langchain-groq>=0.0.1

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Frontend
streamlit>=1.31.0
plotly>=5.18.0

# Database
sqlalchemy>=2.0.0

# Vector Database (RAG)
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Report Generation
fpdf2>=2.7.8
openpyxl>=3.1.2
pandas>=2.1.0

# Web Search
tavily-python>=0.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.0

# Async / Infra
celery>=5.3.6
redis>=5.0.1
aiokafka>=0.10.0